+++ packages/react-workspace/nimble-react/src/anchor-step/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { AnchorStep, anchorStepTag } from '@ni/nimble-components/dist/esm/anchor-step';
 import { wrap } from '../utilities/react-wrapper';