+++ packages/react-workspace/nimble-react/src/breadcrumb/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { Breadcrumb, breadcrumbTag } from '@ni/nimble-components/dist/esm/breadcrumb';
 import { wrap } from '../utilities/react-wrapper';