+++ packages/react-workspace/nimble-react/src/menu-item/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { MenuItem, menuItemTag } from '@ni/nimble-components/dist/esm/menu-item';
 import { wrap, type EventName } from '../utilities/react-wrapper';