+++ packages/react-workspace/nimble-react/src/radio-group/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { RadioGroup, radioGroupTag } from '@ni/nimble-components/dist/esm/radio-group';
 import { wrap, type EventName } from '../utilities/react-wrapper';