+++ packages/react-workspace/ok-react/src/fv/accordion-item/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { FvAccordionItem, fvAccordionItemTag } from '@ni/ok-components/dist/esm/fv/accordion-item';
 import { FvAccordionItemAppearance } from '@ni/ok-components/dist/esm/fv/accordion-item/types';
 import { wrap } from '../../utilities/react-wrapper';